worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael Sambou",email = "msambou@andrew.cmu.edu"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "llama-cpp-python (>=0.3.8,<0.4.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "openai (>=1.75.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[tool.poetry]
packages = [{include = "worker", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
